chore: move stringUtils.ts into a folder /utils
